refactor(dirs): move directories
Put directories back under lib/

#1,#4
