+++ lib/__mocks__/fs-extra.js
