+++ lib/__tests__/base-resource-converter.test.js
@@ -1,11 +1,11 @@
 const allIssuesPrs = require('../__fixtures__/all-issues-prs')
-const baseResourceConverter = require('../lib/base-resource-converter')
-const githubResourceLogger = require('../lib/github-resource-logger')
+const baseResourceConverter = require('../base-resource-converter')
+const githubResourceLogger = require('../github-resource-logger')
 const issuesGetForRepoResponse = require('../__fixtures__/issues/get-for-repo-response')
 const Octokit = require('@octokit/rest')
 const prsGetAllResponse = require('../__fixtures__/pull-requests/get-all-response')
-const defaultOptions = require('../lib/default-options')
-const restClientOptions = require('../lib/rest-client-options')
+const defaultOptions = require('../default-options')
+const restClientOptions = require('../rest-client-options')
 
 describe('baseResourceConverter', () => {
   beforeAll(() => {
