+++ lib/__tests__/file-exporter.test.js
@@ -1,8 +1,8 @@
-const defaultOptions = require('../lib/default-options')
-const exportType = require('../lib/export-type')
-const fileExporter = require('../lib/file-exporter')
+const defaultOptions = require('../default-options')
+const exportType = require('../export-type')
+const fileExporter = require('../file-exporter')
 const fsExtra = require('fs-extra')
-const githubResourceLogger = require('../lib/github-resource-logger')
+const githubResourceLogger = require('../github-resource-logger')
 
 jest.mock('fs-extra')
